Developing an EKF-Slam Implementation for Position Correction
This slam implementation is based on https://github.com/joansola/slamtb one, but I tried to move from a "known-correspondences" implementation to one that does not require a previous knowledge of correspondences between points and landmarks.
To do so, I adapted the original implementation to the approach described in Probabilistic Robotics, Section 10.3.

Further copyright and attribution will be inserted once the implementation is finished, as this is only an ongoing-development commit.